Add ability to execute local applications
HttpActions are now RuleActions as this design has the ability to be very versatile - in this case it now makes sense to have a Rule execute an application on the system, like the system shell, which can run a custom shell script as a provided argument. An 'ExData' field has also been added to the rules.xml for the purpose of sending extra data to a Rule, such as command line args or the body of an HTTP POST request.